requirements.txt
fastapi[standard]
uvicorn
pydantic
langchain
langgraph
langchain-community
langchain-groq
langchain-together
langchain-chroma
langchain-huggingface
transformers
langdetect
tavily-python
chromadb
python-dotenv
langchain-core
rank-bm25
streamlit
httpx
tqdm
